Refactor SignUp page to include account info on success
